event_matcher/requirements.txt
django
firebase-admin
django_bootstrap5
social-auth-app-django
django-extensions
django-widget-tweaks
pymysql
Pillow
django-simple-captcha
google-analytics-data
django-extensions